third_party/WebKit/LayoutTests/editing/spelling/mixed-paste-short.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description('For bug cr571988: Spell checking for pasted mixed content.');

jsTestIsAsync = true;

var testRoot = document.createElement("div");
document.body.insertBefore(testRoot, document.body.firstChild);

var testEditable = document.createElement("div");
testEditable.setAttribute("contentEditable", "true");
testRoot.appendChild(testEditable);

var testSourceDecorated = document.createElement("div");
testSourceDecorated.innerHTML = "<img src='../resources/abe.png'> zz zz.";
testRoot.appendChild(testSourceDecorated);

var sel = window.getSelection();

var tests = [];

function done()
{
    var next = tests.shift();
    if (next)
        return window.setTimeout(next, 0);
    testRoot.style.display = "none";
    finishJSTest();
}

function verifyMarker(node, expectedMarked)
{
    if (node instanceof HTMLInputElement || node instanceof HTMLTextAreaElement) {
        node.focus();
    } else {
        sel.selectAllChildren(node);
    }

    var count = 0;
    for (var i = 0; i < node.childNodes.length; ++i)
        count += internals.markerCountForNode(node.childNodes[i], 'spelling');

    if (count == expectedMarked) {
        var nodeContent = node instanceof HTMLInputElement || node instanceof HTMLTextAreaElement ? node.value : node.innerHTML;
        testPassed(node.tagName + " has " + expectedMarked + " markers on '" + nodeContent + "'");
    }

    return count == expectedMarked;
}

var destination = null;
var misspelledCount = null;
function pasteAndVerify(source, dest, expectedMarked)
{
    sel.selectAllChildren(source);
    document.execCommand("Copy");
    if (dest instanceof HTMLInputElement || dest instanceof HTMLTextAreaElement) {
        dest.value = "";
        dest.focus();
    } else {
        dest.innerHTML = "";
        sel.selectAllChildren(dest);
    }
    document.execCommand("Paste");

    if (window.internals) {
        destination = dest;
        misspelledCount = expectedMarked;
        shouldBecomeEqual('verifyMarker(destination, misspelledCount)', 'true', done);
    }
};

tests.push(function() { pasteAndVerify(testSourceDecorated, testEditable, testSourceDecorated.innerHTML.match(/zz/g).length); });
done();

var successfullyParsed = true;
</script>
</body>
</html>
